first part without tests

+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/web/dto/forclient/dto/LinkUpdateRequest.java
@@ -1,4 +1,4 @@
 package ru.tinkoff.edu.java.scrapper.web.dto.forclient.dto;
 
-public record LinkUpdateRequest(int id, String url, String description, int[] thChatIds) {
+public record LinkUpdateRequest(int id, String url, String description, long[] thChatIds) {
 }
