+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/web/dto/repository/User.java
@@ -1,4 +1,4 @@
 package ru.tinkoff.edu.java.scrapper.web.dto.repository;
 
-public record User(int user_id,int id) {
+public record User(int user_id,long id) {
 }
